refactor revealBitmap and change genMockPassportData path
