+++ common/src/utils/revealBitmap.ts
@@ -23,19 +23,14 @@ export function revealBitmapFromAttributes(attributeToReveal: { [key: string]: b
   return reveal_bitmap;
 }
 
+
 export function unpackReveal(revealedData_packed: string[]): string[] {
-  const revealedData_packed_formatted = [
-    revealedData_packed["revealedData_packed[0]"],
-    revealedData_packed["revealedData_packed[1]"],
-    revealedData_packed["revealedData_packed[2]"],
-  ];
 
   const bytesCount = [31, 31, 28]; // nb of bytes in each of the first three field elements
-  const bytesArray = revealedData_packed_formatted.flatMap((element: string, index: number) => {
+  const bytesArray = revealedData_packed.flatMap((element: string, index: number) => {
     const bytes = bytesCount[index];
     const elementBigInt = BigInt(element);
     const byteMask = BigInt(255); // 0xFF
-
     const bytesOfElement = [...Array(bytes)].map((_, byteIndex) => {
       return (elementBigInt >> (BigInt(byteIndex) * BigInt(8))) & byteMask;
     });
@@ -43,4 +38,14 @@ export function unpackReveal(revealedData_packed: string[]): string[] {
   });
 
   return bytesArray.map((byte: bigint) => String.fromCharCode(Number(byte)));
+}
+
+
+export function formatAndUnpackReveal(revealedData_packed: string[]): string[] {
+  const revealedData_packed_formatted = [
+    revealedData_packed["revealedData_packed[0]"],
+    revealedData_packed["revealedData_packed[1]"],
+    revealedData_packed["revealedData_packed[2]"],
+  ];
+  return unpackReveal(revealedData_packed_formatted);
 }
